# docs/swagger.yaml
definitions:
  dto.SysApiUpdateReq:
    properties:
      apiType:
        description: Path       string `json:"path" comment:"地址"`
        type: string
      title:
        description: Handle     string `json:"handle" comment:"handle"`
        type: string
    type: object
...
info:
  contact: {}
  title: Go-admin 后台管理系统
  version: 2.0.0
paths:
...
  /admin-api/v1/sys-api/{id}:
    put:
      consumes:
      - application/json
      description: 修改接口管理
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysApiUpdateReq'
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      summary: 修改接口管理
      tags:
      - 接口管理
...
swagger: "2.0"